240903웹수정

+++ backend/src/main/java/dw/memorial/Dto/FurnitureDto.java
@@ -16,8 +16,8 @@ public class FurnitureDto {
 
     private String userId;
     private String furnitureObject;
-    private double siteX;
-    private double siteY;
+    private int siteX;
+    private int siteY;
 
 
     public FurnitureDto toGameDtoFromGame(Furniture furniture){
